Move user-agent to constants
